Split out tests to their proper module and do not use test fixtures
