+++ timetable_solutions/tests/unit/domain/data_management/upload_processors/test_lesson.py
@@ -2,68 +2,59 @@
 Module containing unit tests for the LessonFileUploadProcessor
 """
 
-
-# Django imports
-from django import test
+# Third party imports
+import pytest
 
 # Local application imports
 from data import models
 from domain.data_management.upload_processors import LessonFileUploadProcessor
+from tests import data_factories
 
 
-class TestLessonFileUploadProcessor(test.TestCase):
+def _get_file_processor(
+    school: models.School | None = None,
+) -> LessonFileUploadProcessor:
     """
-    Unit test class for the small methods on LessonFileUploadProcessor.
+    LessonFileUploadProcessor that is used throughout the tests.
     """
+    if not school:
+        school = data_factories.School()
 
-    fixtures = [
-        "user_school_profile.json",
-        "year_groups.json",
-        "pupils.json",
-        "timetable.json",
-    ]
+    processor = LessonFileUploadProcessor(
+        csv_file=None,
+        school_access_key=school.school_access_key,
+        attempt_upload=False,
+    )
+    return processor
 
-    @staticmethod
-    def _get_file_processor() -> LessonFileUploadProcessor:
-        """
-        Fixture for a LessonFileUploadProcessor that is used throughout the tests.
-        """
-        # noinspection PyTypeChecker
-        processor = LessonFileUploadProcessor(
-            csv_file=None,
-            school_access_key=123456,
-            attempt_upload=False,
-        )
-        return processor
 
-    # TESTS FOR METHODS GETTING PUPILS / TIMETABLE SLOTS FORM STRING
-    def test_get_pupils_from_raw_pupil_ids_string_valid(self):
-        """
-        Test that when the pupil_ids_raw kwarg is a valid string representing a list of pupils, a queryset
-        of Pupils is returned as expected.
-        """
-        # Set test parameters
-        processor = self._get_file_processor()
+@pytest.mark.django_db
+class TestLessonFileUploadProcessorGetPupilsFromRawPupilIdsString:
+    def test_get_pupils_from_valid_raw_pupil_ids_string(self):
+        school = data_factories.School()
 
-        raw_string = "1, 4, 6"  # Bog-standard string we know should work
+        # Create three pupils to reference the raw ids of
+        factory_pupils = [data_factories.Pupil(school=school) for _ in range(0, 3)]
+        raw_string = "; ".join(str(pupil.pupil_id) for pupil in factory_pupils)
 
-        # Execute test unit
+        # Create some other pupil not expected to be included in the resulting queryset
+        excluded = data_factories.Pupil(school=school)
+
+        processor = _get_file_processor(school=school)
+
+        # Try processing the string of ids to a pupil queryset
         pupils = processor._get_pupils_from_raw_pupil_ids_string(
             pupil_ids_raw=raw_string, row_number=1
         )
 
-        # Check outcomes
-        assert isinstance(pupils, models.PupilQuerySet)
-        assert {pupil.pupil_id for pupil in pupils} == {1, 4, 6}
+        # Ensure all pupils were retrieved
+        assert pupils.count() == 3
+        assert excluded not in pupils
 
-    def test_get_pupils_from_raw_pupil_ids_string_missing_pupils(self):
-        """
-        Test that when a string contains a pupil id without a corresponding Pupil, an error message is set
-        """
-        # Set test parameters
-        processor = self._get_file_processor()
-
-        raw_string = "1, 4, 1000000, 333"  # Note that pupil with ids 1000000 333 does not exist for school 123456
+    def test_errors_for_raw_pupil_ids_string_referencing_non_existent_pupil(self):
+        # Create no pupils, so our raw string reference will be invalid
+        processor = _get_file_processor()
+        raw_string = "1"
 
         # Execute test unit
         pupils = processor._get_pupils_from_raw_pupil_ids_string(
@@ -73,37 +64,37 @@ class TestLessonFileUploadProcessor(test.TestCase):
         # Check outcomes
         assert pupils is None
         assert "No pupil" in processor.upload_error_message
-        assert ("1000000" in processor.upload_error_message) and (
-            "333" in processor.upload_error_message
-        )
 
+
+@pytest.mark.django_db
+class TestLessonFileUploadProcessorGetTimetableSlotsFromRawSlotIdsString:
     def test_get_timetable_slots_from_raw_slot_ids_string_valid(self):
-        """
-        Test that when the slot_ids_raw kwarg is a valid string representing a list of timetable slots, a queryset
-        of TimetableSlots is returned as expected.
-        """
-        # Set test parameters
-        processor = self._get_file_processor()
+        school = data_factories.School()
 
-        raw_string = "1, 2, 3"  # Bog standard string we know should work
+        # Create three pupils to reference the raw ids of
+        factory_slots = [
+            data_factories.TimetableSlot(school=school) for _ in range(0, 3)
+        ]
+        raw_string = "; ".join(str(slot.slot_id) for slot in factory_slots)
 
-        # Execute test unit
+        # Create some other pupil not expected to be included in the resulting queryset
+        excluded = data_factories.TimetableSlot(school=school)
+
+        processor = _get_file_processor(school=school)
+
+        # Try processing the string of ids to a pupil queryset
         slots = processor._get_timetable_slots_from_raw_slot_ids_string(
             slot_ids_raw=raw_string, row_number=1
         )
 
-        # Check outcomes
-        assert isinstance(slots, models.TimetableSlotQuerySet)
-        assert {slot.slot_id for slot in slots} == {1, 2, 3}
+        # Ensure all pupils were retrieved
+        assert slots.count() == 3
+        assert excluded not in slots
 
-    def test_get_timetable_slots_from_raw_slot_ids_string_missing_slots(self):
-        """
-        Test that when a string contains a slot id without a corresponding Timetableslot, an error message is set
-        """
-        # Set test parameters
-        processor = self._get_file_processor()
-
-        raw_string = "1, 2, 123456, 412"  # Note that slots with ids 123456, 412 does not exist for school 123456
+    def test_errors_when_raw_string_references_non_existent_slot(self):
+        # Create no timetable slots, so our raw string reference will be invalid
+        processor = _get_file_processor()
+        raw_string = "1"
 
         # Execute test unit
         slots = processor._get_timetable_slots_from_raw_slot_ids_string(
@@ -113,137 +104,3 @@ class TestLessonFileUploadProcessor(test.TestCase):
         # Check outcomes
         assert slots is None
         assert "No timetable slot" in processor.upload_error_message
-        assert ("123456" in processor.upload_error_message) and (
-            "412" in processor.upload_error_message
-        )
-
-    def test_get_integer_set_from_string_valid_strings(self):
-        """
-        Test that the raw string cleaner is producing the correct set of integers
-        """
-        # Set test parameters
-        processor = self._get_file_processor()
-
-        valid_strings = [
-            "[1, 4, 6]",
-            "1, 4, 6",
-            "1; 4; 6",
-            "1 & 4 & 6",
-            "1, 4; 6 ",
-            "[1 & 4 ; 6]" "[,1, 4, 6",
-            "1, 4, 6]###!!!",
-            "gets-remo, 1, 4, 6, ved ]",
-            "6,,, 4, 1!!!!",
-            "1       ,                4-, 6, ",
-            "[][][1][][], 4[][ ][][, 6[][][][,,,,,,,,,][][][]",
-            "vdwljhcbdwckw1cdwceqdeqdeq][dewqdkebkebd,4,***********\n\n\n\n\n,,,,,,,,new line??? random chars ,,,6",
-        ]
-
-        # Execute test unit (for each string, effectively redo the test)
-        for valid_string in valid_strings:
-            integer_set = processor.get_integer_set_from_string(
-                raw_string_of_ids=valid_string, row_number=1
-            )
-
-            # Check outcome
-            assert integer_set == {
-                1,
-                4,
-                6,
-            }, f"{valid_string} caused test failure with wrong integer set: {integer_set}"
-
-    def test_get_integer_set_from_string_invalid_harmless_strings_return_none(self):
-        """
-        Test that the raw string cleaner is correctly handling strings that can signal an empty queryset
-        """
-        # Set test parameters
-        processor = self._get_file_processor()
-
-        invalid_strings = [
-            "",
-            "[]",
-            "[][][][][][]",
-            "None",
-            "N/A",
-            "no slots / no pupils",
-        ]
-
-        # Execute test unit (for each string, effectively redo the test)
-        for invalid_string in invalid_strings:
-            integer_set = processor.get_integer_set_from_string(
-                raw_string_of_ids=invalid_string, row_number=1
-            )
-
-            # Check outcome
-            assert integer_set is None, f"{invalid_string} caused test failure!"
-
-    # TESTS FOR STRING CLEANING ON INDIVIDUAL IDS
-    def test_get_clean_id_from_file_field_value_unchanged_for_integers(self):
-        """
-        Test that when we pass an INTEGER id (the ideal format), it is unchanged by the cleaner.
-        """
-        # Set tst parameters
-        raw_ids = [1, 10, 100, 123, 423412]
-        processor = self._get_file_processor()
-
-        # Execute test unit
-        for raw_id in raw_ids:
-            cleaned_id = processor.get_clean_id_from_file_field_value(
-                user_input_id=raw_id, row_number=1, field_name="irrelevant"
-            )
-
-            # Check outcome
-            assert cleaned_id == raw_id
-
-    def test_get_clean_id_from_file_field_value_changed_for_floats(self):
-        """
-        Test that when we pass an FLOAT id the 0s don't get counted as factors of 10.
-        """
-        # Set tst parameters
-        raw_ids = [1.0, 10.0, 100.0, 123.0, 423412.0]
-        expected_converted_ids = [1, 10, 100, 123, 423412]
-        processor = self._get_file_processor()
-
-        # Execute test unit
-        for n, raw_id in enumerate(raw_ids):
-            cleaned_id = processor.get_clean_id_from_file_field_value(
-                user_input_id=raw_id, row_number=1, field_name="irrelevant"
-            )
-
-            # Check outcome
-            assert cleaned_id == expected_converted_ids[n]
-
-    def test_get_clean_id_from_file_field_value_changed_for_strings(self):
-        """
-        Test that when we pass strings as IDs, they are handled in the appropriate way.
-        """
-        # Set tst parameters
-        raw_ids = ["1.0", "10.0", "100", "123.0", "423412.0"]
-        expected_converted_ids = [1, 10, 100, 123, 423412]
-        processor = self._get_file_processor()
-
-        # Execute test unit
-        for n, raw_id in enumerate(raw_ids):
-            cleaned_id = processor.get_clean_id_from_file_field_value(
-                user_input_id=raw_id, row_number=1, field_name="irrelevant"
-            )
-
-            # Check outcome
-            assert cleaned_id == expected_converted_ids[n]
-
-    def test_get_clean_id_from_file_field_error_for_string_list(self):
-        """
-        Test that when we pass strings as IDs, they are handled in the appropriate way.
-        """
-        # Set tst parameters
-        raw_id = "1, 2, 3"
-        processor = self._get_file_processor()
-
-        # Execute test unit
-        processor.get_clean_id_from_file_field_value(
-            user_input_id=raw_id, row_number=1, field_name="irrelevant"
-        )
-
-        # Check outcome
-        assert processor.upload_error_message is not None
-        assert "Multiple" in processor.upload_error_message
